moved SuspendResume class to os/posix directory
